requirements.txt
ultralytics>=8.2.0
opencv-python-headless>=4.9.0
streamlit>=1.28.0
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0